Refactor release_drafter.yml and main.py
