+++ main.py
@@ -1,10 +1,13 @@
+import sys
 import time
 from typing import Union
 import getpass
 
+from hydra import compose, initialize
 import logfire
 from adbutils import AdbDevice
 from pydantic import computed_field, model_validator
+from omegaconf import OmegaConf
 import pyautogui
 from src.compare import ImageComparison
 from src.get_screen import GetScreen
@@ -97,9 +100,15 @@ class RemoteContoller(ConfigModel):
             time.sleep(self.global_interval)
 
 
-if __name__ == "__main__":
-    from src.utils.config_utils import load_hydra_config
+def load_hydra_config() -> dict:
+    args = sys.argv[1:]
+    with initialize(config_path="./configs", version_base="1.3"):
+        cfg = compose(config_name="configs", overrides=args, return_hydra_config=False)
+        config_dict = OmegaConf.to_container(cfg, resolve=False)
+    return config_dict
 
+
+if __name__ == "__main__":
     configs = load_hydra_config()
     game_config = configs["games"]
     auto_web = RemoteContoller(**game_config)
