+++ pyproject.toml
@@ -67,6 +67,7 @@ dependencies = [
     "hydra-core>=1.3.2",
     "hydra-colorlog>=1.2.0",
     "hydra-optuna-sweeper>=1.2.0",
+    "rootutils>=1.0.7",
 ]
 readme = "README.md"
 requires-python = ">= 3.10"
