+++ requirements-dev.lock
@@ -269,6 +269,7 @@ python-dateutil==2.9.0.post0
 python-dotenv==1.0.1
     # via autorootcwd
     # via pyrootutils
+    # via rootutils
 pytweening==1.2.0
     # via pyautogui
 pywin32==306
@@ -291,6 +292,8 @@ retry==0.9.2
     # via adbutils
 rich==13.7.1
     # via logfire
+rootutils==1.0.7
+    # via auto-click
 scipy==1.14.0
     # via optuna
 setuptools==71.0.3
