+++ src/compare.py
@@ -9,6 +9,7 @@ import numpy as np
 import logfire
 from pydantic import Field, BaseModel, ConfigDict, computed_field, model_validator
 import PIL.Image as Image
+
 from src.types.image_models import ImageModel
