+++ src/types/config.py
@@ -1,6 +1,7 @@
 from typing import Optional
 
 from pydantic import Field, BaseModel
+
 from src.types.image_models import ImageModel
